+++ src/Components/Common/Card/Card.tsx
@@ -1,21 +1,10 @@
 import { Button, Input } from '..';
-import {
-  useCardSide,
-  useInput,
-  useAtomInput,
-  useCorrect,
-} from '../../../hooks';
-import {
-  AnswerContainer,
-  CardBackContainer,
-  CardWrapper,
-  Paragraph,
-  CardEditContainer,
-} from './Card.style';
+import { useCardSide, useInput } from '../../../hooks';
+import { CardWrapper, CardEditContainer } from './Card.style';
 import { useMutation } from '@tanstack/react-query';
 import { updateCardsAPI } from '../../../api/cardClient';
-import { CardSetting } from './CardSetting';
 import { CardFront } from './CardFront';
+import { CardBack } from './CardBack';
 
 export function Card({ question, answer, _id, stackCount }: Card) {
   return (
@@ -38,52 +27,6 @@ export function Card({ question, answer, _id, stackCount }: Card) {
         </>
       )}
     </CardWrapper>
-  );
-}
-
-type CardBackProps = {
-  _id: string;
-  answer: string;
-  question: string;
-  stackCount: number;
-};
-
-function CardBack({ _id, answer, question, stackCount }: CardBackProps) {
-  const { mutate: updateCard } = useMutation({ mutationFn: updateCardsAPI });
-  const { cardSide, toggleFront } = useCardSide();
-
-  const { isCorrect } = useCorrect();
-  const { inputVal, resetInputVal } = useAtomInput();
-
-  const handleConform = () => {
-    resetInputVal();
-    toggleFront();
-
-    if (_id) {
-      const submitDate = new Date();
-      if (isCorrect) {
-        updateCard({
-          id: _id,
-          card: { question, answer, submitDate, stackCount: stackCount + 1 },
-        });
-      } else {
-        updateCard({
-          id: _id,
-          card: { question, answer, submitDate, stackCount: 0 },
-        });
-      }
-    }
-  };
-
-  return (
-    <CardBackContainer active={cardSide === 'back'} isCorrect={isCorrect}>
-      <CardSetting _id={_id} />
-      <AnswerContainer>
-        <Paragraph>정답: {answer}</Paragraph>
-        <Paragraph>풀이: {inputVal}</Paragraph>
-      </AnswerContainer>
-      <Button onClick={handleConform}>확인</Button>
-    </CardBackContainer>
   );
 }
