+++ src/Components/Common/Card/CardBack/index.ts
@@ -1,0 +1,1 @@
+export * from './CardBack';
